+++ clang/docs/ReleaseNotes.md
@@ -54,6 +54,13 @@ in a future version of Clang.
   mode, as it was removed from the standard by
   [P3475R2](https://wg21.link/P3475R2).
 
+- An error diagnostic is now issued if an `#embed` directive or a `__has_embed`
+  expression contains a parameter (i.e. `limit`, `prefix`, `suffix`, `if_empty`)
+  that has previously been defined as a macro, as per
+  [CWG3013](https://wg21.link/cwg3013). Previously macros that shared names with
+  `#embed` parameter names were expanded regardless, similar to its behavior in
+  C: Note that this expansion behavior is still present in C since there are no
+  rule analogue to CWG3013 in C.
 
 ### Objective-C Specific Potentially Breaking Changes
 
@@ -520,6 +527,12 @@ features cannot lower the translation-unit ABI level;
   keyword of an alias-declaration. (#GH155787)
 
 - Improve Clang diagnoses when unary `__imag` operator with non-complex type operand is used as lvalue. (GH222383)
+
+- Added `-Wembed-parameter-is-macro`, which warns in C if an `#embed` directive
+  or a `__has_embed` expression uses a parameter (i.e. `limit`, `prefix`,
+  `suffix`, `if_empty`) that has also been defined as a macro. C expands the
+  macro, but the same code is ill-formed in C++, so `-Wembed-parameter-is-macro`
+  is also part of `-Wc++-compat`; this warning is disabled by default otherwise.
 
 ### Improvements to Clang's time-trace